few parts refactored
